Hot fix, Plataforma adaptada
