+++ lib/logg.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+import 'package:jobjenius/navegacion.dart';
+
+
+
+class Main extends StatelessWidget {
+  const Main({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const MaterialApp(
+      debugShowCheckedModeBanner: false,
+      home: Navigation(),
+    );
+  }
+}
