+++ lib/navegacion.dart
@@ -1,7 +1,8 @@
 import 'package:curved_navigation_bar/curved_navigation_bar.dart';
 import 'package:flutter/material.dart';
-import 'package:jobjenius/Busqueda.dart';
+import 'package:jobjenius/BusquedaTrabajadores.dart';
 import 'package:jobjenius/Trabajadores.dart';
+import 'package:jobjenius/perfilCliente.dart';
 import 'package:jobjenius/seguridad.dart';
 import 'theme/app_color.dart';
 class Navigation extends StatefulWidget {
@@ -22,7 +23,7 @@ class _NavigationState extends State<Navigation> {
     BusquedaTabajadores(),
     ModuloInicio(),
     Seguridad(),
-    ModuloInicio(),
+    Perfilcliente(),
 
   ];
